Use Gate in ProductController (allows)

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreProductRequest;
 use App\Http\Requests\UpdateProductRequest;
 use App\Models\Product;
+use Illuminate\Support\Facades\Gate;
 
 class ProductController extends Controller
 {
@@ -22,7 +23,9 @@ class ProductController extends Controller
      */
     public function create()
     {
-        return view('products.create');
+        if (Gate::allows('product-manage'))
+            return view('products.create');
+        abort(403);
     }
 
     /**
@@ -30,8 +33,11 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
-        Product::create($request->all());
-        return redirect()->route('products.index');
+        if (Gate::allows('product-manage')){
+            Product::create($request->all());
+            return redirect()->route('products.index');
+        }
+        abort(403);
     }
 
 
@@ -40,7 +46,9 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
-        return view('products.edit', compact('product'));
+        if (Gate::allows('product-manage'))
+            return view('products.edit', compact('product'));
+        abort(403);
     }
 
     /**
@@ -48,8 +56,11 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, Product $product)
     {
-        $product->update($request->all());
-        return redirect()->route('products.index');
+        if (Gate::allows('product-manage')){
+            $product->update($request->all());
+            return redirect()->route('products.index');
+        }
+        abort(403);
     }
 
     /**
@@ -57,7 +68,10 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        $product->delete();
-        return redirect()->route('products.index');
+        if (Gate::allows('product-delete')) {
+            $product->delete();
+            return redirect()->route('products.index');
+        }
+        abort(403);
     }
 }
